+++ built_json/test/built_list_serializer_test.dart
@@ -89,7 +89,7 @@ void main() {
 
   group('BuiltList with unknown specifiedType and no builders', () {
     final data = new BuiltList<int>([1, 2, 3]);
-    final specifiedType = const FullType();
+    final specifiedType = FullType.unspecified;
     final serializers = new Serializers();
     final serialized = [
       'list',