+++ built_json/test/string_serializer_test.dart
@@ -26,7 +26,7 @@ void main() {
   group('String with unknown specifiedType', () {
     final data = 'testing, testing';
     final serialized = ['String', 'testing, testing'];
-    final specifiedType = const FullType();
+    final specifiedType = FullType.unspecified;
 
     test('can be serialized', () {
       expect(serializers.serialize(data, specifiedType: specifiedType), serialized);