+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Stop falling back to "<Object>" for collection type parameters.
 - Extend code generation to add the builders needed for collections.
+- Differentiate Object from unspecified.
 
 ## 0.0.2